rename: ChatMessageResponseDto 애플리케이션 디렉토리 변경
