+++ chat/src/main/java/com/team15gijo/chat/domain/model/v1/ChatMessageDocument.java
@@ -1,6 +1,6 @@
 package com.team15gijo.chat.domain.model.v1;
 
-import com.team15gijo.chat.presentation.dto.v1.ChatMessageResponseDto;
+import com.team15gijo.chat.application.dto.v1.ChatMessageResponseDto;
 import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import java.time.LocalDateTime;
